requirements.txt
numpy 1.24.3
pandas 1.5.3
streamlit 1.27.2
Pillow 9.4.0
tensorflow 2.14.0
scikit-learn 1.2.2